requirements.txt
flask
discord.py
dotenv
